move parse_args to x84/cmdline.py
