+++ x84/msgpoll.py
@@ -1,54 +1,17 @@
 #!/usr/bin/env python2.7
-"""
-x84net message poll for x/84, https://github.com/jquast/x84
+""" x84net message poll for x/84. """
 
-To configure message polling, add a tag for the network to the ``'server_tags'``
-attribute in the ``[msg]`` section of your default.ini.  Optionally include a
-custom 'origin' line.  If not provided, ``"Sent from <bbsname>"`` will be
-used.
-
-Next, create a section using the name of that tag, prefixed with
-``'msgnet_'``.  For example, if the tag is ``'x84net'``, create a
-``'msgnet_x84net'`` section.
-
-The following attributes are required:
-
- - ``url_base``: The base URL for the message network's REST API.
- - ``board_id``: Your board's ID in the network.
- - ``token``: Your board's secure token, assigned to you by the network admin.
-
-The following attributes are optional:
-
- - ``ca_path``: The path to a CA bundle if the server's CA is not already
-   included in your operating system.
- - ``poll_interval``: The number of seconds elapsed between polling a message
-   network for new messages (default is 1984, ~33 minutes).
-
-If you wish to tag your messages with a custom origin line when they are
-delivered to the network hub, add an 'origin_line' attribute to the ``[msg]``
-section of your ``default.ini``.
-
-Example *default.ini* configuration::
-
-    [msg]
-    network_tags = x84net
-    origin_line = Sent from a mediocre BBS.
-
-    [msgnet_x84net]
-    url_base = https://some.server:8443/api/
-    board_id = 1
-    token = somereallylongtoken
-    poll_interval = 300
-"""
-
-# local imports
+# std imports
 import logging
 import hashlib
 import time
 import json
 import os
 
-# 3rd party imports
+# local
+from . import cmdline
+
+# 3rd-party
 import requests
 
 
@@ -413,7 +376,7 @@ if __name__ == '__main__':
     # of web modules and ssl options may be gathered.
     import x84.engine
     import x84.bbs.ini
-    x84.bbs.ini.init(*x84.engine.parse_args())
+    x84.bbs.ini.init(*cmdline.parse_args())
 
     # do not execute message polling as a background thread.
     main(background_daemon=False)
